Fix Slack Bolt migration
